+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-fee.spec.ts
@@ -1,6 +1,6 @@
 import {
   activityDummyData,
-  symbolProfileDummyData,
+  assetProfileDummyData,
   userDummyData
 } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator-test-utils';
 import { PortfolioCalculatorFactory } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator.factory';
@@ -85,17 +85,17 @@ describe('PortfolioCalculator', () => {
       const activities: Activity[] = [
         {
           ...activityDummyData,
-          date: new Date('2021-09-01'),
-          feeInAssetProfileCurrency: 49,
-          feeInBaseCurrency: 49,
-          quantity: 0,
-          SymbolProfile: {
-            ...symbolProfileDummyData,
+          assetProfile: {
+            ...assetProfileDummyData,
             currency: 'USD',
             dataSource: 'MANUAL',
             name: 'Account Opening Fee',
             symbol: '2c463fb3-af07-486e-adb0-8301b3d72141'
           },
+          date: new Date('2021-09-01'),
+          feeInAssetProfileCurrency: 49,
+          feeInBaseCurrency: 49,
+          quantity: 0,
           type: 'FEE',
           unitPriceInAssetProfileCurrency: 0
         }